+++ seqr/views/apis/igv_api.py
@@ -2,12 +2,121 @@ import re
 from django.http import StreamingHttpResponse
 from django.contrib.auth.decorators import login_required
 
+from seqr.models import Individual, IgvSample
+from seqr.views.utils.dataset_utils import validate_alignment_dataset_path # TODO wrong file?
 from seqr.utils.file_utils import file_iter, does_file_exist
-from seqr.views.utils.permissions_utils import get_project_and_check_permissions
+from seqr.views.utils.json_utils import create_json_response
+from seqr.views.utils.orm_to_json_utils import  get_json_for_sample
+from seqr.views.utils.permissions_utils import get_project_and_check_permissions, check_project_permissions, \
+    data_manager_required
 from settings import API_LOGIN_REQUIRED_URL
 
 import logging
 logger = logging.getLogger(__name__)
+
+
+@data_manager_required
+def receive_igv_table_handler(request, project_guid):
+    project = get_project_and_check_permissions(project_guid, request.user, can_edit=True)
+    info = []
+
+    def _process_alignment_records(rows, **kwargs):
+        invalid_row = next((row for row in rows if not 2 <= len(row) <= 3), None)
+        if invalid_row:
+            raise ValueError("Must contain 2 or 3 columns: " + ', '.join(invalid_row))
+        parsed_records = defaultdict(list)
+        for row in rows:
+            parsed_records[row[0]].append({'filePath': row[1], 'sampleId': row[2] if len(row)> 2 else None})
+        return parsed_records
+
+    try:
+        uploaded_file_id, filename, individual_dataset_mapping = save_uploaded_file(request, process_records=_process_alignment_records)
+
+        matched_individuals = Individual.objects.filter(family__project=project, individual_id__in=individual_dataset_mapping.keys())
+        unmatched_individuals = set(individual_dataset_mapping.keys()) - {i.individual_id for i in matched_individuals}
+        if len(unmatched_individuals) > 0:
+            raise Exception('The following Individual IDs do not exist: {}'.format(", ".join(unmatched_individuals)))
+
+        info.append('Parsed {} rows in {} individuals from {}'.format(
+            sum([len(rows) for rows in individual_dataset_mapping.values()]), len(individual_dataset_mapping), filename))
+
+        existing_sample_files = defaultdict(set)
+        for sample in IgvSample.objects.select_related('individual').filter(individual__in=matched_individuals):
+            existing_sample_files[sample.individual.individual_id].add(sample.file_path)
+
+        unchanged_rows = set()
+        for individual_id, updates in individual_dataset_mapping.items():
+            unchanged_rows.update([
+                (individual_id, update['filePath']) for update in updates
+                if update['filePath'] in existing_sample_files[individual_id]
+            ])
+
+        if unchanged_rows:
+            info.append('No change detected for {} rows'.format(len(unchanged_rows)))
+
+        all_updates = []
+        for i in matched_individuals:
+            all_updates += [
+                dict(individualGuid=i.guid, **update) for update in individual_dataset_mapping[i.individual_id]
+                if (i.individual_id, update['filePath']) not in unchanged_rows
+            ]
+
+    except Exception as e:
+        return create_json_response({'errors': [str(e)]}, status=400)
+
+    response = {
+        'updates': all_updates,
+        'uploadedFileId': uploaded_file_id,
+        'errors': [],
+        'info': info,
+    }
+    return create_json_response(response)
+
+
+SAMPLE_TYPE_MAP = {
+    'bam': IgvSample.SAMPLE_TYPE_ALIGNMENT,
+    'cram': IgvSample.SAMPLE_TYPE_ALIGNMENT,
+    'bigWig': IgvSample.SAMPLE_TYPE_COVERAGE,
+    'junctions.bed.gz': IgvSample.SAMPLE_TYPE_JUNCTION,
+    'dcr.bed.gz': IgvSample.SAMPLE_TYPE_GCNV,
+}
+
+@data_manager_required
+def update_individual_igv_sample(request, individual_guid):
+    individual = Individual.objects.get(guid=individual_guid)
+    project = individual.family.project
+    check_project_permissions(project, request.user, can_edit=True)
+
+    request_json = json.loads(request.body)
+
+    try:
+        file_path = request_json.get('filePath')
+        if not file_path:
+            raise ValueError('request must contain fields: filePath')
+
+        suffix = '.'.join(file_path.split('.')[1:])
+        sample_type = SAMPLE_TYPE_MAP.get(suffix)
+        if not sample_type:
+            raise Exception('Invalid file extension for "{}" - valid extensions are {}'.format(
+                file_path, ', '.join(SAMPLE_TYPE_MAP.keys())))
+        validate_alignment_dataset_path(file_path)
+
+        sample, created = get_or_create_model_from_json(
+            IgvSample, create_json={'individual': individual, 'sample_type': sample_type},
+            update_json={'file_path': file_path, 'sample_id': request_json.get('sampleId')}, user=request.user)
+
+        response = {
+            'igvSamplesByGuid': {
+                sample.guid: get_json_for_sample(sample, individual_guid=individual_guid, project_guid=project.guid)}
+        }
+        if created:
+            response['individualsByGuid'] = {
+                individual.guid: {'igvSampleGuids': [s.guid for s in individual.igvsample_set.all()]}
+            }
+        return create_json_response(response)
+    except Exception as e:
+        error = str(e)
+        return create_json_response({'error': error}, status=400, reason=error)
 
 
 @login_required(login_url=API_LOGIN_REQUIRED_URL)
